refactor: inline provider routines into api app
