--- packages/provider-routines/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@repo/provider-routines",
-  "version": "0.1.0",
-  "private": true,
-  "license": "Apache-2.0",
-  "type": "module",
-  "sideEffects": false,
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "default": "./src/index.ts"
-    }
-  },
-  "scripts": {
-    "clean": "git clean -xdf .cache .turbo node_modules",
-    "test": "vitest run",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@db/app": "workspace:*",
-    "@lightfast/connector-core": "workspace:*",
-    "@lightfast/connector-linear": "workspace:*",
-    "@repo/api-contract": "workspace:*",
-    "@repo/app-encryption": "workspace:*"
-  },
-  "devDependencies": {
-    "@repo/typescript-config": "workspace:*",
-    "@repo/vitest-config": "workspace:*",
-    "@types/node": "catalog:",
-    "typescript": "catalog:",
-    "vitest": "catalog:"
-  }
-}
