--- packages/provider-routines/vitest.config.ts
@@ -1,12 +1,0 @@
-import sharedConfig from "@repo/vitest-config";
-import { defineConfig, mergeConfig } from "vitest/config";
-
-export default mergeConfig(
-  sharedConfig,
-  defineConfig({
-    test: {
-      globals: true,
-      environment: "node",
-    },
-  })
-);
